Mini projet fonctionnel

+++ src/abstractTree/IEnvironment.java
@@ -1,9 +1,9 @@
 package abstractTree;
 
-public interface IEnvironment {
-	
-	// Retrouve une variable d'un environnement
-	Integer getVariableValue(String variable);
-	
-	void putVariable(String var, Types type, Integer value);
+public interface IEnvironment
+{
+    // Retrouve une variable d'un environnement
+    Double getVariableValue(String variable);
+
+    void putVariable(String var, Types type, Double value);
 }
